requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-google-genai
google-generativeai
langchain-community
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
streamlit
python-dotenv
pydantic
python-dateutil
pytz
httpx
requests
structlog
parsedatetime 